Make paths friendlier, func. descriptions, etc
The os.path library seems to know how to differentiate \ and / for Windows/Linux. Hereon out going to be implementing that wherever possible. Also added some function/class descriptions but not all of them yet, and some clarifications as to what is going on where. Removed the "cache_throw_disabled" and put more specific exceptions into each spot it was in. Completely altered/cleaned description parsing in courses, but still needs work.