v3 beginning
There has been a newfound interest in this project, so I'm returning to it.. I found a PHP page that responds to specific queries that the script will be redesigned around - as it uses XML there should be some good libraries for handling that sort of thing. Can't wait to get into it.