Second pass uses existing downloaded pages
Commit title is the main optimization, but lots more happened behind the scenes. Imports were adjusted for the hierarchy changes in this branch. A few functions needed to be split up into two lines, as one-lining it wasn't working for some reason. Removed the requirement for term for now, as only the current year's programs will be supported for now. further optimized compute and readability, and most importantly: it runs, and gives I believe the same output as the original code - all bachelor and associates degrees. Grad degrees are not possible.