Re-added terms
Added term back in as the basis for the accordion function to get its classes. Also prints a line saying what term it's getting courses from - by default, the latest, but can be passed to the degrees method. Also added to the to-do a point to implement accessing catalogs older than/other than the current one - likely by a given term's academic year e.g. Fall 2023 and Spring 2024 would both want info from the 2023-2024 catalog, but not Spring 2023.